Hide unguarded API when lock is enabled
- Conditionally disable doc & integration tests if lock is enabled as
  they would break otherwise
- Use `doc_cfg` to document SpiceLock on docs.rs
