+++ rust-spice/src/lib.rs
@@ -1,8 +1,9 @@
-#![doc = include_str!("../README.md")]
 #![doc(
     html_logo_url = "https://raw.githubusercontent.com/GregoireHENRY/rust-spice/main/rust-spice/rsc/img/logo_squared.png",
     html_favicon_url = "https://raw.githubusercontent.com/GregoireHENRY/rust-spice/main/rust-spice/rsc/img/logo_squared.png"
 )]
+// Doc tests in the README fail with 'lock' enabled
+#![cfg_attr(not(feature = "lock"), doc = include_str!("../README.md"))]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
 #[cfg(feature = "default")]
@@ -26,14 +27,21 @@ pub mod c {
     */
     pub use cspice_sys::*;
 }
+
+// The unguarded API should only be exposed if the lock is disabled
+#[cfg(not(feature = "lock"))]
 pub mod core;
-
-// TODO: We probably want to break the usual API when the lock is enabled to prevent accidental
-// usage, but that breaks integration and doc tests
-// #[cfg(any(test, not(feature = "lock")))]
+#[cfg(not(feature = "lock"))]
 pub use crate::core::*;
-
-pub use crate::core::{DLADSC, DSKDSC};
-
+// If it is enabled, only this crate should see it
 #[cfg(feature = "lock")]
+pub(crate) mod core;
+#[cfg(feature = "lock")]
+pub(crate) use crate::core::*;
+
+// These items need to be exposed regardless of whether 'lock' is enabled or not
+pub use crate::core::{DLADSC, DSKDSC, MAX_LEN_OUT, TIME_FORMAT, TIME_FORMAT_SIZE};
+
+#[cfg(any(feature = "lock", doc))]
+#[cfg_attr(docsrs, doc(cfg(feature = "lock")))]
 pub use crate::core::lock::SpiceLock;
